+++ bc/family_information/tests/fixtures.py
@@ -4,11 +4,12 @@ import wagtail_factories
 from bc.family_information import models
 
 
-class FamilyInformationHomePageFactory(wagtail_factories.PageFactory):
+class SubsiteHomePageFactory(wagtail_factories.PageFactory):
     class Meta:
-        model = models.FamilyInformationHomePage
+        model = models.SubsiteHomePage
 
     title = factory.Sequence(lambda n: f"Familiy Information Service Home Page {n}")
+    listing_summary = "Familiy Information Service Home Page"
 
     banner_image = factory.SubFactory("bc.images.tests.fixtures.ImageFactory")
     banner_title = "Banner title"
@@ -23,6 +24,7 @@ class CategoryTypeOnePageFactory(wagtail_factories.PageFactory):
         model = models.CategoryTypeOnePage
 
     title = factory.Sequence(lambda n: f"Category Type One Page {n}")
+    listing_summary = "Category Type One Page"
 
     banner_image = factory.SubFactory("bc.images.tests.fixtures.ImageFactory")
     banner_title = "Banner title"