+++ bc/inlineindex/tests/fixtures.py
@@ -9,6 +9,7 @@ class InlineIndexFactory(wagtail_factories.PageFactory):
         model = models.InlineIndex
 
     title = factory.Sequence(lambda n: f"InlineIndex {n}")
+    listing_summary = "Inline Index Summary"
 
 
 class InlineIndexChildFactory(wagtail_factories.PageFactory):
@@ -16,3 +17,4 @@ class InlineIndexChildFactory(wagtail_factories.PageFactory):
         model = models.InlineIndexChild
 
     title = factory.Sequence(lambda n: f"InlineIndexChild {n}")
+    listing_summary = "Inline Index Child Summary"