+++ bc/recruitment/tests/fixtures.py
@@ -30,6 +30,7 @@ class RecruitmentHomePageFactory(factory.django.DjangoModelFactory):
         model = "recruitment.RecruitmentHomePage"
 
     title = factory.Sequence(lambda n: f"Recruitment HomePage")
+    listing_summary = "Recruitment HomePage"
     hero_image = factory.SubFactory("bc.images.tests.fixtures.ImageFactory")
     hero_title = "foo"
     hero_link_text = "foo"